Move humidifier consts to separate module
